Add function association for default cache behavior in CloudFront. (#5)
As it turns out, navigating this type of site works just fine in the
browser, but if you go to share a link (and thus navigate to a page
directly), there is no referrer information to make sure that the user
ends up on the right page, and they end up getting the `404` page.

This adds a function that interpolates the uri to add `.html` if the
request itself is missing referrer information, so that we still
navigate to the right page. Works fine for this use-case.

The `terraform plan` should show no changes, since I applied the changes
locally for testing.